fix(Tree): allows catch all to be used as optional globbing
Catch all parameter was limited to be used *after* a slash (`/`) on
a path (ie. `/members/*trailing`).

Attempts to use it immediately after the path (ie. `/members*trailing`)
was not properly identified or captured.

This change ensures catch all can be used as globbing, allowing capture
of anything after the matching path.

Closes #12